+++ src/Presentation/Nop.Web/Factories/IPollModelFactory.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using Nop.Core.Domain.Polls;
+using Nop.Web.Models.Polls;
+
+namespace Nop.Web.Factories
+{
+    public partial interface IPollModelFactory
+    {
+        PollModel PreparePollModel(Poll poll, bool setAlreadyVotedProperty);
+
+        PollModel PreparePollModelBySystemName(string systemKeyword);
+
+        List<PollModel> PrepareHomePagePollModels();
+    }
+}
